+++ source/extensions/stat_sinks/metrics_service/grpc_metrics_service_impl.h
@@ -112,25 +112,9 @@ public:
 
   void beginFlush() override { message_.clear_envoy_metrics(); }
 
-  void flushCounter(const Stats::Counter& counter, uint64_t) override {
-    io::prometheus::client::MetricFamily* metrics_family = message_.add_envoy_metrics();
-    metrics_family->set_type(io::prometheus::client::MetricType::COUNTER);
-    metrics_family->set_name(counter.name());
-    auto* metric = metrics_family->add_metric();
-    metric->set_timestamp_ms(std::chrono::system_clock::now().time_since_epoch().count());
-    auto* counter_metric = metric->mutable_counter();
-    counter_metric->set_value(counter.value());
-  }
-
-  void flushGauge(const Stats::Gauge& gauge, uint64_t value) override {
-    io::prometheus::client::MetricFamily* metrics_family = message_.add_envoy_metrics();
-    metrics_family->set_type(io::prometheus::client::MetricType::GAUGE);
-    metrics_family->set_name(gauge.name());
-    auto* metric = metrics_family->add_metric();
-    metric->set_timestamp_ms(std::chrono::system_clock::now().time_since_epoch().count());
-    auto* gauage_metric = metric->mutable_gauge();
-    gauage_metric->set_value(value);
-  }
+  void flushCounter(const Stats::Counter& counter, uint64_t) override;
+  void flushGauge(const Stats::Gauge& gauge, uint64_t value) override;
+  void flushHistogram(const Stats::ParentHistogram& histogram) override;
 
   void endFlush() override {
     grpc_metrics_streamer_->send(message_);
@@ -140,9 +124,7 @@ public:
     }
   }
 
-  void onHistogramComplete(const Stats::Histogram&, uint64_t) override {
-    // TODO : Need to figure out how to map existing histogram to Proto Model
-  }
+  void onHistogramComplete(const Stats::Histogram&, uint64_t) override {}
 
 private:
   GrpcMetricsStreamerSharedPtr grpc_metrics_streamer_;